Fix AutoTab bug in vimrc
The function assumes that the cursor is at the start of the buffer, and
runs on the FileType event. Change this to run on BufRead.

Also set the cursor position to the top of the file before searching for
indentation. This may be unnecessary, with the previous change, but it
might be, and it shouldn't hurt.