Add quotation marks around files in Makefile
When running in Cygwin on Windows, we need to account for the fact that
Windows allows user names to have spaces, so the shell expands ~ to have
a space.

The rules that execute ~/.vim-plugins/update and ~/.zsh-plugins/update
may not work on Windows still, but quotation marks won't be able to help
since the shell doesn't expand "~".