Don't define command_not_found_handle() in bashrc
The idea behind it was to save space by compressing binaries and only
extracting them as necessary, but in practice that level of difference
in space in negligible, and defining a custom handle for when a command
isn't found can cause several problems.

In addition, I never made nor really wanted to get this to work for zsh.