Create transfer target and worker script
`make transfer` will create compressed archives that can be ssh'd over
to another system and then uncompressed to transfer over the necessary files
to transfer settings from one system to another. By default, it creates
a .zip file (for portability), and a .tar.xz file (for transferring less
data).

A separate script is used to do the actual work to keep the makefile
clean and avoid issues with recursive make and dependencies.