Update ssh-agent starting
This will allow for when there is no common parent terminal. Currently,
every new terminal emulator (except the first) finds that ssh-agent is
already running, but it doesn't have the environment variables necessary
to send to ssh commands. This commit will make it so that every
ssh-agent binds to the same file.

The filename is just the SHA1 hash of "joel-ssh-agent".