Fix update scripts bug; use master, not last tag
Add 1 to denominator to prevent division by zero when there aren't many
plugins.

The tag change only really was relevant for 1 issue in 1 plugin, so
that's probably better as a special case if it comes up again. Use
master instead, and clean (git wipe) the directory before working with
it.